+++ lib/common/utils/iterable_utils.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/widgets.dart';
+
 /// Extensions on Iterable class.
 extension IterableUtils<T> on Iterable<T> {
 
@@ -11,5 +13,22 @@ extension IterableUtils<T> on Iterable<T> {
   /// and cast the resulting Iterable<T?> to Iterable<T>.
   Iterable<T> filterNotNull() {
     return where((e) => e != null).cast();
+  }
+}
+
+/// Extensions on Iterable with Widgets.
+extension IterableWidgets<T> on Iterable<Widget> {
+
+  /// Creates a new iterable with widgets, which are separated via given
+  /// [element] separator widget.
+  Iterable<Widget> separated(Widget element) sync* {
+    final iterator = this.iterator;
+    if (iterator.moveNext()) {
+      yield iterator.current;
+      while(iterator.moveNext()) {
+        yield element;
+        yield iterator.current;
+      }
+    }
   }
 }