+++ lib/models/dto/contact_dto.g.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'contact_dto.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+ContactDto _$ContactDtoFromJson(Map<String, dynamic> json) => ContactDto(
+      name: json['name'] as String,
+      phoneNumber: json['phoneNumber'] as String?,
+    );
+
+Map<String, dynamic> _$ContactDtoToJson(ContactDto instance) =>
+    <String, dynamic>{
+      'name': instance.name,
+      'phoneNumber': instance.phoneNumber,
+    };